Added more paramters to GenFeatureSet.
All maps can now be created individually by using the corresponding parameters. Max/Min/Non Crossing maps are now optional. If no parameter is used, all non optional maps are generated.

Number of processes can now be set with a parameter. Default is 16 or the number of processors on the system if not enough cores are available.