Added a second tqdm dialog showing the progress.
The second tqdm dialog is only present in SLIXParameterGenerator because
single line profiles get evaluated way too quickly for tqdm to be
useful. Therefore it will only be implemented there.

Reimplemented the smoothing algorithm present in previous versions of
SLIX. The base parameters might be wrong in the current state. The
method also needs to be tested for high coverage!

Signed-off-by: Jan Andre Reuter <j.reuter@fz-juelich.de>