Changes to direction
Previously when the distance between two corresponding peaks were outside of the range of ]145°, 215°[ the direction would not be calculated for that singular peak. However it is more safe to assume that the whole line profile cannot be used to calculate the direction.
With this commit line profiles with a peak distance outside of this range will return BACKGROUND_COLOR for all directions. You would get something like [BACKGROUND_COLOR, BACKGROUND_COLOR, 60] before. This will be [BACKGROUND_COLOR, BACKGROUND_COLOR, BACKGROUND_COLOR] now.

Signed-off-by: Jan André Reuter <jan.andre.reuter@hotmail.de>