Added first version of weighting vectors maps through parameter maps.
The current implementation results in some unexpected results, because Matplotlib seems to scale all vectors based on the range of the vector lengths. Because of that, the vectors will be shown in a smaller scale than expected. To circumvent this tissue, one can try to use the --scale option with double the --thinout value to get an accurate representation

Signed-off-by: Jan André Reuter <jan.andre.reuter@hotmail.de>